ループ変数の再利用を実装し、ループ変数のインデックス型をu32に変更

+++ src/ir/builder.rs
@@ -136,7 +136,7 @@ impl Builder {
         });
     }
 
-    pub fn push_loop(&mut self, loop_var: usize, breakable: Breakable) {
+    pub fn push_loop(&mut self, loop_var: u32, breakable: Breakable) {
         self.push(Inst {
             kind: InstKind::Loop(loop_var),
             breakable,
